+++ MaterialDesignExtensionsDemo/ViewModel/AutocompleteViewModel.cs
@@ -103,4 +103,46 @@ namespace MaterialDesignExtensionsDemo.ViewModel
             return m_operatingSystemItems.Where(item => item.Name.ToLower().Contains(searchTerm));
         }
     }
+
+    // using an abstract class instead of the interface
+    //
+    //public class OperatingSystemAutocompleteSource : AutocompleteSource<OperatingSystemItem>
+    // or
+    //public class OperatingSystemAutocompleteSource : AutocompleteSourceChangingItems<OperatingSystemItem>
+    //{
+    //    private List<OperatingSystemItem> m_operatingSystemItems;
+
+    //    public OperatingSystemAutocompleteSource()
+    //    {
+    //        m_operatingSystemItems = new List<OperatingSystemItem>()
+    //        {
+    //            new OperatingSystemItem() { Icon = PackIconKind.Android, Name = "Android Gingerbread" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Android, Name = "Android Icecream Sandwich" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Android, Name = "Android Jellybean" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Android, Name = "Android Lollipop" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Android, Name = "Android Nougat" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Linux, Name = "Debian" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.DesktopMac, Name = "Mac OSX" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Raspberrypi, Name = "Raspbian" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Ubuntu, Name = "Ubuntu Wily Werewolf" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Ubuntu, Name = "Ubuntu Xenial Xerus" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Ubuntu, Name = "Ubuntu Yakkety Yak" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Ubuntu, Name = "Ubuntu Zesty Zapus" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Windows, Name = "Windows 7" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Windows, Name = "Windows 8" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Windows, Name = "Windows 8.1" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Windows, Name = "Windows 10" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Windows, Name = "Windows Vista" },
+    //            new OperatingSystemItem() { Icon = PackIconKind.Windows, Name = "Windows XP" }
+    //        };
+    //    }
+
+    //    public override IEnumerable<OperatingSystemItem> Search(string searchTerm)
+    //    {
+    //        searchTerm = searchTerm ?? string.Empty;
+    //        searchTerm = searchTerm.ToLower();
+
+    //        return m_operatingSystemItems.Where(item => item.Name.ToLower().Contains(searchTerm));
+    //    }
+    //}
 }